+++ tests/Type/PollAnswerTest.php
@@ -15,10 +15,11 @@ final class PollAnswerTest extends TestCase
 {
     public function testBase(): void
     {
-        $answer = new PollAnswer('sadg', [2, 4]);
+        $answer = new PollAnswer('sadg', [2, 4], ['pid2', 'pid4']);
 
         assertSame('sadg', $answer->pollId);
         assertSame([2, 4], $answer->optionIds);
+        assertSame(['pid2', 'pid4'], $answer->optionPersistentIds);
         assertNull($answer->voterChat);
         assertNull($answer->user);
     }
@@ -28,6 +29,7 @@ final class PollAnswerTest extends TestCase
         $answer = (new ObjectFactory())->create([
             'poll_id' => 'sadg',
             'option_ids' => [2, 4],
+            'option_persistent_ids' => ['pid2', 'pid4'],
             'voter_chat' => [
                 'id' => 42,
                 'type' => 'private',
@@ -41,6 +43,7 @@ final class PollAnswerTest extends TestCase
 
         assertSame('sadg', $answer->pollId);
         assertSame([2, 4], $answer->optionIds);
+        assertSame(['pid2', 'pid4'], $answer->optionPersistentIds);
         assertSame(42, $answer->voterChat?->id);
         assertSame(43, $answer->user?->id);
     }