+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - New #191: Add `allowsRevoting`, `shuffleOptions`, `allowAddingOptions`, `hideResultsUntilCloses`, `description`, 
   `descriptionParseMode` and `descriptionEntities` parameters to `SendPoll` method.
 - New #191: Add `persistentId`, `addedByUser`, `addedByChat` and `additionDate` fields to `PollOption` type.
+- New #191: Add `optionPersistentIds` field to `PollAnswer` type.
 
 ## 0.17 March 1, 2026